feat: Update 2.25.4 migration

+++ packages/snjs/lib/migrations/2_25_4.ts
@@ -19,13 +19,18 @@ export class Migration2_25_4 extends Migration {
     const themes = this.services.itemManager.getItems(
       ContentType.Theme
     ) as SNTheme[];
-    const noDistractionIdentifier = 'org.standardnotes.theme-no-distraction' as FeatureIdentifier;
+    const noDistractionIdentifier =
+      'org.standardnotes.theme-no-distraction' as FeatureIdentifier;
     const noDistractionTheme = themes.find(
       (theme) => theme.package_info.identifier === noDistractionIdentifier
     ) as SNItem;
-    this.services.itemManager.removeItemLocally(noDistractionTheme);
-    await this.services.storageService.deletePayloadWithId(
-      noDistractionTheme.uuid
-    );
+    if (noDistractionTheme) {
+      await this.services.itemManager.setItemToBeDeleted(
+        noDistractionTheme.uuid
+      );
+      await this.services.storageService.deletePayloadWithId(
+        noDistractionTheme.uuid
+      );
+    }
   }
 }
